design: 홈 화면과 시작하기 화면 통합 완료
