+++ src/app/(home)/_components/HeroSection.tsx
@@ -1,12 +1,9 @@
 'use client';
 
 import { useModal } from '@/hooks/useModal';
-import { bannersData } from '@/mocks/home/index';
-
 import styled from 'styled-components';
 import Title from '@/components/common/Title';
 import Button from '@/components/common/Button';
-import BannerSlider from '@/components/common/slider/BannerSlider';
 import Trial from '@/app/(home)/_components/Trial';
 
 export default function HeroSection() {
@@ -18,18 +15,19 @@ export default function HeroSection() {
 
 	return (
 		<StyledHeroSection>
-			<div className="left-section">
-				<div className="description">
-					<Title size="small">새로운 AI 구독 서비스</Title>
-					<Title size="large" weight="semiBold">
-						원하는 분야만 골라보는 나만의 뉴스
-					</Title>
+			<div className="content">
+				<div className="left-section">
+					<div className="description">
+						<Title size="small">새로운 AI 구독 서비스</Title>
+						<Title size="large" weight="semiBold">
+							원하는 분야만 골라보는 나만의 뉴스
+						</Title>
+					</div>
+					<Button size="large" scheme="primary" style={{ marginTop: '1rem' }} onClick={handleOpenModal}>
+						AI 뉴스레터 체험하기
+					</Button>
 				</div>
-				<Button size="large" scheme="primary" style={{ marginTop: '1rem' }} onClick={handleOpenModal}>
-					AI 뉴스레터 체험하기
-				</Button>
 			</div>
-			<BannerSlider banners={bannersData} />
 		</StyledHeroSection>
 	);
 }
@@ -37,32 +35,56 @@ export default function HeroSection() {
 const StyledHeroSection = styled.section`
 	width: 100%;
 	height: fit-content;
-	display: flex;
 
-	flex-direction: row;
+	display: flex;
+	flex-direction: column;
 	justify-content: space-between;
 	align-items: center;
+
 	gap: 2rem;
 	padding: 1rem 0 2rem 0;
+	margin-bottom: 4rem;
 
-	.left-section {
+	.header {
 		width: 100%;
-		height: 100%;
-
 		display: flex;
-		flex-wrap: wrap;
 		flex-direction: column;
-		justify-content: space-between;
+		justify-content: flex-start;
 		align-items: flex-start;
-		padding: 2rem 0;
 	}
 
-	img {
-		border-radius: ${({ theme }) => theme.borderRadius.soft};
-		box-shadow: ${({ theme }) => theme.shadow.medium};
-	}
+	.content {
+		width: 100%;
+		height: fit-content;
+		display: flex;
+		flex-direction: row;
+		justify-content: center;
+		align-items: center;
 
-	@media ${({ theme }) => theme.mediaQuery.tablet} {
-		flex-wrap: wrap;
+		.left-section {
+			width: 100%;
+			height: 100%;
+
+			display: flex;
+			flex-wrap: wrap;
+			flex-direction: column;
+			justify-content: space-between;
+			align-items: center;
+			padding: 2rem 0;
+
+			.description {
+				text-align: center;
+				align-items: center;
+			}
+		}
+
+		img {
+			border-radius: ${({ theme }) => theme.borderRadius.soft};
+			box-shadow: ${({ theme }) => theme.shadow.medium};
+		}
+
+		@media ${({ theme }) => theme.mediaQuery.tablet} {
+			flex-wrap: wrap;
+		}
 	}
 `;
