+++ src/app/(home)/_components/SubscribeSection.tsx
@@ -17,6 +17,7 @@ import { BiCheck, BiPlus } from 'react-icons/bi';
 import { LuMailCheck } from 'react-icons/lu';
 import { useModal } from '@/hooks/useModal';
 import ArgreementCheck from '@/components/common/article/AgreementCheck';
+import FullWidthPanel from '@/components/common/FullWidthPanel';
 
 interface Props {
 	trends: IArticleSummary[];
@@ -51,64 +52,72 @@ const SubscribeSection = ({ trends }: Props) => {
 
 	return (
 		<StyledSubscribe>
-			<CardSlider
-				className="quick-subscription"
-				type="sub"
-				data={CATEGORIES.map((category) => {
-					// 카테고리 이름이 '전체'인 경우 기본 이미지를 사용.
-					const image =
-						category.name === '전체'
-							? '/img/category_all-2.jpg'
-							: trends.find((trend) => trend.categoryName === category.name)?.image ||
-							  '/img/newpick_default_img.jpg';
-
-					return {
-						id: category.id ?? 0,
-						url: `/articles` + (category.id !== 0 ? `?categoryId=${category.id}` : ''),
-						image,
-						header: category.name,
-						main: {
-							title: undefined,
-							description: `${category.name} 분야의 최신 뉴스레터를 구독하세요.`,
-						},
-						footer: (
-							<Button
-								key={category.id}
-								scheme={selectedInterests.includes(category.name) ? 'primary' : 'outline'}
-								onClick={() => handleSelectInterests(category)}
-								icon={selectedInterests.includes(category.name) ? <BiCheck /> : <BiPlus />}
-								style={{
-									width: '100%',
-								}}
-								disabled={isChangingSubscription}
-							>
-								{selectedInterests.includes(category.name) ? <>Selected</> : <>Select</>}
-							</Button>
-						),
-					};
-				})}
-			/>
-
-			<div className="subscription-form">
-				<Title size="medium" weight="bold">
-					주요 분야 빠른 구독하기
+			<FullWidthPanel>
+				<Title size="extraLarge" weight="bold" color="background">
+					📩 지금 뉴스레터를 시작해보세요
 				</Title>
-				<Text size="small">선택한 분야를 이메일로 보내드릴게요.</Text>
-				<div className="main">
-					<Button
-						type="submit"
-						scheme="primary"
-						style={{
-							width: '100%',
-							marginTop: '0.5rem',
-						}}
-						onClick={handleSubscribe}
-						disabled={isChangingSubscription}
-					>
-						구독 신청
-					</Button>
+			</FullWidthPanel>
+
+			<div className="content">
+				<CardSlider
+					className="quick-subscription"
+					type="sub"
+					data={CATEGORIES.map((category) => {
+						// 카테고리 이름이 '전체'인 경우 기본 이미지를 사용.
+						const image =
+							category.name === '전체'
+								? '/img/category_all-2.jpg'
+								: trends.find((trend) => trend.categoryName === category.name)?.image ||
+								  '/img/newpick_default_img.jpg';
+
+						return {
+							id: category.id ?? 0,
+							url: `/articles` + (category.id !== 0 ? `?categoryId=${category.id}` : ''),
+							image,
+							header: category.name,
+							main: {
+								title: undefined,
+								description: `${category.name} 분야의 최신 뉴스레터를 구독하세요.`,
+							},
+							footer: (
+								<Button
+									key={category.id}
+									scheme={selectedInterests.includes(category.name) ? 'primary' : 'outline'}
+									onClick={() => handleSelectInterests(category)}
+									icon={selectedInterests.includes(category.name) ? <BiCheck /> : <BiPlus />}
+									style={{
+										width: '100%',
+									}}
+									disabled={isChangingSubscription}
+								>
+									{selectedInterests.includes(category.name) ? <>Selected</> : <>Select</>}
+								</Button>
+							),
+						};
+					})}
+				/>
+
+				<div className="subscription-form">
+					<Title size="medium" weight="bold">
+						주요 분야 빠른 구독하기
+					</Title>
+					<Text size="small">선택한 분야를 이메일로 보내드릴게요.</Text>
+					<div className="main">
+						<Button
+							type="submit"
+							scheme="primary"
+							style={{
+								width: '100%',
+								marginTop: '0.5rem',
+							}}
+							onClick={handleSubscribe}
+							disabled={isChangingSubscription}
+						>
+							구독 신청
+						</Button>
+					</div>
+					<ArgreementCheck name={checkName} />
 				</div>
-				<ArgreementCheck name={checkName} />
 			</div>
 		</StyledSubscribe>
 	);
@@ -117,8 +126,16 @@ const SubscribeSection = ({ trends }: Props) => {
 const StyledSubscribe = styled.section`
 	width: 100%;
 	display: flex;
-	gap: 2rem;
-	margin: 4rem auto;
+	flex-direction: column;
+	margin-top: 6rem;
+	margin-bottom: 4rem;
+
+	.content {
+		display: flex;
+		flex-direction: row;
+		justify-content: space-between;
+		gap: 2rem;
+	}
 
 	.subscription-form {
 		width: fit-content;
